+++ tests/fixtures/project_s/project_s/__init__.py
@@ -1,0 +1,5 @@
+"""Top-level package for project_n."""
+
+__author__ = """full_n"""
+__email__ = 'email_'
+__version__ = '1.0'