+++ tests/fixtures/project_yehua/project_yehua/__init__.py
@@ -1,0 +1,3 @@
+# flake8: noqa
+from project_yehua._version import __version__
+from project_yehua._version import __author__