+++ tests/fixtures/project_yehua/test.sh
@@ -1,0 +1,2 @@
+pip freeze
+nosetests --with-coverage --cover-package project_yehua --cover-package tests tests  docs/source project_yehua