+++ tests/test_main.py
@@ -8,17 +8,19 @@ from mock import patch
 from nose.tools import eq_, raises
 
 
-@patch("os.system")
+@patch("yehua.project.os.chdir")
+@patch("yehua.project._run_command")
 @patch("yehua.project.get_user_inputs")
 @patch("yehua.utils.mkdir")
 @patch("yehua.utils.save_file")
 @patch("yehua.utils.copy_file")
-def test_main(copy, save, mkdir, inputs, os_system):
+def test_main(copy, save, mkdir, inputs, *_):
     copy.return_value = 0
     save.return_value = 0
     mkdir.return_value = 0
     inputs.return_value = dict(project_name="test-me")
-    main()
+    with patch.object(sys, "argv", ["yh"]):
+        main()
     calls = mkdir.call_args_list
     calls = [str(call) for call in calls]
     expected = [
@@ -33,15 +35,6 @@ def test_main(copy, save, mkdir, inputs, os_system):
         "call('test-me/.moban.d/docs/source')",
     ]
     eq_(calls, expected)
-
-
-@raises(SystemExit)
-def test_main_help():
-    args = ["yehua", "help"]
-    with patch("sys.stdout", new_callable=StringIO) as out:
-        with patch.object(sys, "argv", args):
-            main()
-            eq_(out.getvalue(), HELP_TEXT)
 
 
 @raises(SystemExit)