+++ yehua/project.py
@@ -1,16 +1,21 @@
 import os
+import subprocess
 from datetime import datetime
 
 import yehua.utils as utils
+from yehua.utils import get_user_inputs
+from moban.externals.file_system import exists, is_dir, read_unicode
 
-from jinja2 import Environment, FileSystemLoader
+import fs
+from jinja2 import Environment
+from jinja2_fsloader import FSLoader
 
 
 class Project:
     def __init__(self, yehua_file):
-        if not os.path.exists(yehua_file):
+        if not exists(yehua_file):
             raise Exception("%s does not exist" % yehua_file)
-        if os.path.isdir("yehua_file"):
+        if is_dir(yehua_file):
             raise Exception("A yehua file is expected. Not a directory")
 
         self.project_file = yehua_file
@@ -37,6 +42,8 @@ class Project:
                 utils.save_file(target, rendered_content)
 
     def copy_static_files(self):
+        if "static" not in self.directives:
+            return
         for static in self.directives["static"]:
             for output, source in static.items():
                 source = os.path.abspath(os.path.join(self.static_dir, source))
@@ -44,8 +51,15 @@ class Project:
                 utils.copy_file(source, dest)
 
     def inflate_all_by_moban(self):
-        cmd = "cd %s && moban" % self.answers["project_name"]
-        os.system(cmd)
+        current = os.getcwd()
+        project_name = self.answers["project_name"]
+        os.chdir(project_name)
+        cmd = "moban"
+        _run_command(cmd)
+        os.chdir(current)
+        utils.color_print(
+            f"\u2713 Files are generated under [info]{project_name}[/info]"
+        )
 
     def post_moban(self):
         if "post-moban" not in self.directives:
@@ -56,28 +70,42 @@ class Project:
 
     def initialize_git_and_add_all(self, project_files):
         project_name = self.answers["project_name"]
-        cmd = "cd %s && git init" % project_name
-        os.system(cmd)
+        current = os.getcwd()
+        os.chdir(project_name)
+        cmd = "git init"
+        _run_command(cmd)
         for file_name in project_files:
-            cmd = "cd %s && git add %s" % (project_name, file_name)
-            os.system(cmd)
-        print("Please review changes before commit!")
+            _run_command(f"git add {file_name}")
+        os.chdir(current)
+        utils.color_print(
+            f"\u2713 Git repo initialized under [info]{project_name}[/info]"
+            + " and is ready to commit"
+        )
 
     def end(self):
-        print("All done!! project %s is created" % self.project_name)
+        utils.color_print(
+            "All done!! project [info]%s[/info] is created."
+            % self.project_name
+        )
+        utils.color_print(
+            "In the future, "
+            + "run [info]moban[/info] to synchronize with the project template"
+        )
 
     def _ask_questions(self):
-        base_path = os.path.dirname(self.project_file)
-        with open(self.project_file, "r") as f:
-            first_stage = utils.load_yaml(f.read())
-            print(first_stage["introduction"])
+        content = read_unicode(self.project_file)
+        first_stage = utils.load_yaml(content)
+        utils.color_print(first_stage["introduction"])
+        base_path = fs.path.dirname(self.project_file)
+        with fs.open_fs(base_path) as the_fs:
             self.template_dir = os.path.join(
-                base_path, first_stage["configuration"]["template_path"]
+                the_fs._root_path,
+                first_stage["configuration"]["template_path"],
             )
             self.static_dir = os.path.join(
-                base_path, first_stage["configuration"]["static_path"]
+                the_fs._root_path, first_stage["configuration"]["static_path"]
             )
-            self.answers = get_user_inputs(first_stage["questions"])
+        self.answers = get_user_inputs(first_stage["questions"])
 
     def _append_magic_variables(self):
         self.project_name = self.answers["project_name"]
@@ -86,51 +114,35 @@ class Project:
         self.jj2_environment = self._create_jj2_environment(self.template_dir)
 
     def _template_yehua_file(self):
-        base_path = os.path.dirname(self.project_file)
-        tmp_env = self._create_jj2_environment(base_path)
-        template = tmp_env.get_template(os.path.basename(self.project_file))
-        renderred_content = template.render(**self.answers)
-        self.directives = utils.load_yaml(renderred_content)
+        base_path = fs.path.dirname(self.project_file)
+        with fs.open_fs(base_path) as the_fs:
+            base_path = the_fs._root_path
+            tmp_env = self._create_jj2_environment(base_path)
+            template = tmp_env.get_template(
+                os.path.basename(self.project_file)
+            )
+            renderred_content = template.render(**self.answers)
+            self.directives = utils.load_yaml(renderred_content)
 
     def _create_jj2_environment(self, path):
-        template_loader = FileSystemLoader(path)
+        template_loader = FSLoader(path)
+        default_extensions = [
+            "cookiecutter.extensions.JsonifyExtension",
+            "jinja2_time.TimeExtension",
+        ]
         environment = Environment(
             loader=template_loader,
             keep_trailing_newline=True,
             trim_blocks=True,
             lstrip_blocks=True,
+            extensions=default_extensions,
         )
         return environment
 
 
-def get_user_inputs(questions):
-    answers = {}
-    for q in questions:
-        for key, question in q.items():
-            if isinstance(question, list):
-                q, additional = raise_complex_question(question)
-                answers[key] = q
-                if additional:
-                    answers.update(additional)
-            else:
-                a = utils.yehua_input(question + " ")
-                answers[key] = a
-    return answers
-
-
-def raise_complex_question(question):
-    additional_answers = None
-    for subq in question:
-        subquestion = subq.pop("question")
-        suggested_answers = sorted(subq.keys())
-        long_question = [subquestion] + suggested_answers
-        choice = "(%s): " % (
-            ",".join([str(x) for x in range(1, len(long_question))])
-        )
-        long_question.append(choice)
-        a = utils.yehua_input("\n".join(long_question))
-        for key in suggested_answers:
-            if key.startswith(a) and subq[key] != "N/A":
-                additional_answers = get_user_inputs(subq[key])
-        break
-    return a, additional_answers
+def _run_command(command):
+    subprocess.check_call(
+        command.split(" "),
+        stdout=subprocess.DEVNULL,
+        stderr=subprocess.DEVNULL,
+    )